requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.0.0

# Google Gemini SDK
google-genai>=0.3.0

# Vector Database
pinecone-client>=3.0.0

# Embeddings (torch installed separately in Dockerfile)
sentence-transformers>=2.2.0

# Data Processing
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# CLI
typer>=0.9.0
rich>=13.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0